requirements.txt
Flask
spotipy
python-dotenv
transformers
torch
